pyproject.toml
[project]
name = "genai-agents"
version = "0.1.0"
description = "Curso de agentes de IA generativa con LangChain para desarrolladores"
authors = [
    { name = "Alan Sastre" }
]
readme = "README.md"
requires-python = ">=3.13.5"
license = { text = "MIT" }
keywords = ["langchain", "agents", "ai", "llm", "rag", "langgraph"]
dependencies = [
    "langchain>=1.2.10",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "langchain-tavily>=0.2.17",
    "langgraph-checkpoint-postgres>=3.0.4",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langsmith>=0.7.6",
    "python-dotenv>=1.2.1",
]
